+++ wp-parsely/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.20.7](https://github.com/Parsely/wp-parsely/compare/3.20.6...3.20.7) - 2025-09-09
+
+### Fixed
+
+- Engagement Boost: Make preview work for non-administrators ([#3636](https://github.com/Parsely/wp-parsely/pull/3636))
+- Allow full metadata in homepages ([#3571](https://github.com/Parsely/wp-parsely/pull/3571))
+
+### Dependency Updates
+
+- The list of all dependency updates for this release is available [here](https://github.com/Parsely/wp-parsely/pulls?q=is%3Apr+is%3Amerged+milestone%3A3.20.7+label%3ADeps).
+
 ## [3.20.6](https://github.com/Parsely/wp-parsely/compare/3.20.5...3.20.6) - 2025-07-22
 
 ### Dependency Updates